Feat: 서치바 컴포넌트화
